Created KeyboardSizeListener
