+++ Quiz Challenge/Quiz Challenge/Controllers/QuizViewController.swift
@@ -11,31 +11,17 @@ import UIKit
 class QuizViewController: UIViewController {
 
     @IBOutlet weak var bottomPanelPositionConstraint: NSLayoutConstraint!
+    var keyboardSizeListener: KeyboardSizeListener!
 
     override func viewDidLoad() {
         super.viewDidLoad()
 
         // Do any additional setup after loading the view.
-        let notificationCenter = NotificationCenter.default
-        notificationCenter.addObserver(self, selector: #selector(onKeyboardChangeNotification), name: UIResponder.keyboardWillHideNotification, object: nil)
-        notificationCenter.addObserver(self, selector: #selector(onKeyboardChangeNotification), name: UIResponder.keyboardWillChangeFrameNotification, object: nil)
+        keyboardSizeListener = KeyboardSizeListener(self, selector: #selector(onKeyboardUpdate))
     }
 
-    @objc func onKeyboardChangeNotification(notification: Notification) {
-        guard
-            let frameData = notification.userInfo?[UIResponder.keyboardFrameEndUserInfoKey],
-            let keyboardScreenEndFrame = (frameData as? NSValue)?.cgRectValue
-        else {
-            return
-        }
+    @objc func onKeyboardUpdate(height: CGFloat) {
 
-        let keyboardViewEndFrame = view.convert(keyboardScreenEndFrame, from: view.window)
-        let keyboardHeight = keyboardViewEndFrame.height - view.safeAreaInsets.bottom
-
-        self.onKeyboardUpdate(height: keyboardHeight)
-    }
-
-    func onKeyboardUpdate(height: CGFloat) {
         bottomPanelPositionConstraint.constant = height
         view.layoutIfNeeded()
     }
